+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - update libcosmic
 - add pages
+- vendor wl-clipboard code
 
 ## [0.1.0] - 2024-11-19